chore(directories): move cli to top level
