+++ cli/index.js
